\name{getgrib-package}
\alias{getgrib-package}
\alias{getgrib}
\docType{package}
\title{\packageTitle{getgrib}}
\description{\packageDescription{getgrib}}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{getgrib}
\packageIndices{getgrib}

The package provides some helper functions to extract data from grib1
files. Originally developed to provide a fortran/R interface for nearest
neighbor data extraction from grib1 files using the ECMWF grib_api.

Required: EMCWF grib_api installed.
}
\author{
\packageAuthor{getgrib}
Maintainer: \packageMaintainer{getgrib}
}
%%\references{}
\keyword{ package }
%\seealso{
%}
%\examples{
%}
